# Conflicts:
#	earnit.iml
#	src/main/java/nl/earnit/dao/DAOManager.java
#	src/main/java/nl/earnit/dao/WorkedWeekDAO.java
#	src/main/java/nl/earnit/models/db/WorkedWeek.java

+++ src/main/java/nl/earnit/models/db/UserContract.java
@@ -8,12 +8,12 @@ import java.math.BigInteger;
 public class UserContract {
     private String id;
     private String contractId;
-    private BigInteger hourlyWage;
+    private Float hourlyWage;
     private boolean active;
 
     public UserContract() {}
 
-    public UserContract(String id, String contractId, BigInteger hourlyWage, boolean active) {
+    public UserContract(String id, String contractId, Float hourlyWage, boolean active) {
         this.id = id;
         this.contractId = contractId;
         this.hourlyWage = hourlyWage;
@@ -36,11 +36,11 @@ public class UserContract {
         this.contractId = contractId;
     }
 
-    public BigInteger getHourlyWage() {
+    public Float getHourlyWage() {
         return hourlyWage;
     }
 
-    public void setHourlyWage(BigInteger hourlyWage) {
+    public void setHourlyWage(Float hourlyWage) {
         this.hourlyWage = hourlyWage;
     }
